howtodev-server/requirements.txt
channels==2.4.0
channels-redis==2.4.2
Django==3.0.3
django-rq==2.3.0
djangorestframework==3.11.0
psycopg2-binary==2.8.4
pytest-asyncio==0.10.0
pytest-django==3.8.0
django-cors-headers
pytest==5.4.1
pytest-asyncio==0.10.0
pytest-cov==2.8.1
pytest-django==3.8.0
pytest-factoryboy==2.0.3
pytest-pythonpath==0.7.3
pytest-sugar==0.9.3
pytest-testdox==1.2.1
